+++ src/pages/connectYanoljaAccountPage/components/accountVerificationSection/AccountVerificationSection.style.ts
