+++ src/pages/connectYanoljaAccountPage/components/termsAgreementSection/TermsAgreementSection.style.ts
